+++ test/API/ParticleSystemPositions.jl
@@ -146,7 +146,7 @@
     @test length(take!(io)) > 0
 
     # error on construction if dimension is not inferrable
-    @test_throws "method matching length" ParticleSystemPositions(Vector{Float64}[])
+    @test_throws BoundsError ParticleSystemPositions(Vector{Float64}[])
 
 end