+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Version 0.10.0-DEV
 - ![BREAKING][badge-breaking] The previous "lower level interface" is now internal and not exported or public anymore.
 - ![BREAKING][badge-breaking] The `autoswap` function was removed, in favor of the new `pairwise!(f, x, sys)` method. It not available anymore in for `neighborlist` functions. 
 - ![BREAKING][badge-breaking] The python interface was currently discontinued.
+- ![BREAKING][badge-breaking] `validate_coordinates` has to be a `Function` (`(x)  -> nothing` to skip all validation).
 - ![INFO][badge-info] `AbstractParticleSystem`, `ParticleSystem1` and `ParticleSystem2` are documented as public, to control dispatch in advanced applications.
 - ![INFO][badge-info] Renamed internal methods of `pairwise!` to `_pairwise!` to improve error messages. 
 - ![INFO][badge-info] Remove custom linear algebra code, use Base methods.